airflow/requirements.txt
apache-airflow-providers-postgres==5.11.0
apache-airflow-providers-http==4.10.0
faker==26.0.0
httpx==0.27.2
psycopg2-binary==2.9.9
